+++ warp/__init__.py
@@ -393,6 +393,13 @@ from warp._src.utils import TIMING_GRAPH as TIMING_GRAPH
 from warp._src.utils import TIMING_ALL as TIMING_ALL
 
 
+# category: Profiling > CUDA Profiler
+
+from warp._src.context import cuda_profiler_start as cuda_profiler_start
+from warp._src.context import cuda_profiler_stop as cuda_profiler_stop
+from warp._src.context import cuda_profiler_range as cuda_profiler_range
+
+
 # category: Logging
 
 from warp._src.logger import LOG_DEBUG as LOG_DEBUG